The dwellposition, catheter, and catheter table are in seperate files
